pipeline/requirements.txt
celery[redis]==5.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.0
pymupdf==1.24.0
feedparser==6.0.11
boto3==1.34.0
anthropic==0.40.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.0
pytest-httpx==0.30.0
